Fix building on linux
In order for INT64_MIN and INT64_MAX to be defined, __STDC_LIMIT_MACROS must
be defined.

Linking would fail when using "ld", due to "undefined reference to __dso_handle"
so instead, use "c++" to link.

When linking, if "-undefined dynamic_lookup" was passed as a parameter to c++
the link would fail due to being unable to find "dynamic_lookup" (ie, the option
of "-undefined" wasn't properly parsed).  Only add that option in the "darwin"
case rather than having it apply for all cases.